Replace use of Q with native Promise.
Now that we required Node.js 4.x or above, we've got native Promise support.
These changes removes the use of the 3rd party promise library Q,
as we don't have much use for extra promise goodies in this package,
and it's seen to be better to be a lean package than having this dependency.

BREAKING CHANGE as the returned promises from plex-api instances will now
be *native* Promise instances, rather than Q promise instances.

Refs the first Q -> native Promise change: https://github.com/phillipj/node-plex-api/commit/2ea75ce3f7d0d3c7e25f5d4768b1eb976edb687f